+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+owncloud-spreedme (0.1.6)
+* Fix for ownCloud 9, bug was introduced in the previous release
+
+owncloud-spreedme (0.1.5)
+* Allow non-ownCloud admins to generate 'Temporary Passwords' if they are admin for a 'Spreed.ME' group
+* Various other changes
+
 owncloud-spreedme (0.1.4)
 * External users (invited via 'Temporary Password') are now marked as such
 * Various other changes